Refactored pages structure and removed DocView error
