+++ src/components/DocsView/pages/CommunityPage.tsx
