+++ src/components/DocsView/pages/ResourcesPage.tsx
