+++ src/main.tsx
@@ -8,11 +8,11 @@ import AboutView from './components/AboutView/AboutView'
 import DocsView from './components/DocsView/DocsView'
 import FAQ from './components/DocsView/components/FAQ'
 import * as FastClick from 'fastclick'
-import QuickstartPage from './components/DocsView/pages/quickstart/QuickstartPage'
-import ReferencePage from './components/DocsView/pages/reference/ReferencePage'
-import ResourcesPage from './components/DocsView/pages/resources/ResourcesPage'
-import BlogPage from './components/DocsView/pages/blog/BlogPage'
-import CommunityPage from './components/DocsView/pages/community/CommunityPage'
+import QuickstartPage from './components/DocsView/pages/QuickstartPage'
+import ReferencePage from './components/DocsView/pages/ReferencePage'
+import ResourcesPage from './components/DocsView/pages/ResourcesPage'
+import BlogPage from './components/DocsView/pages/BlogPage'
+import CommunityPage from './components/DocsView/pages/CommunityPage'
 import ContentHandler from './components/DocsView/components/ContentHandler'
 import ApolloClient, {createNetworkInterface} from 'apollo-client'
 import { ApolloProvider } from 'react-apollo'
